requirements.txt
matplotlib~=3.5.2

easyocr~=1.5.0
Pillow~=9.1.1
numpy~=1.22.4
imageio~=2.19.3